requirements.txt
pillow>=4.2.1
django-taggit>=0.22.2
django-ipware>=1.1.6
requests>(unknown for now)